refact: alter value to float into schema

+++ api/database/migrations/1552437622916_product_schema.js
@@ -12,7 +12,7 @@ class ProductSchema extends Schema {
         .references('id')
         .inTable('categories')
         .notNullable()
-      table.integer('value').notNullable()
+      table.float('value').notNullable()
       table.integer('amount').notNullable()
       table.increments()
       table.timestamps()
